+++ lib/canary/utils.ex
@@ -103,7 +103,7 @@ defmodule Canary.Utils do
   """
   @spec required?(Keyword.t()) :: boolean
   def required?(opts) do
-    !!Keyword.get(opts, :required, false)
+    !!Keyword.get(opts, :required, true)
   end
 
   @doc """
@@ -135,4 +135,97 @@ defmodule Canary.Utils do
     end
   end
 
+  @doc """
+  Get the resource name from the options, convert it to an atom, and pluralize
+  it when needed — only for the `:index` action.
+
+  If the `:as` option is provided, it is used as the resource name.
+
+      iex> Canary.Utils.get_resource_name(:show, model: MyApp.Post)
+      :post
+
+      iex> Canary.Utils.get_resource_name(:index, model: MyApp.Post)
+      :posts
+
+      iex> Canary.Utils.get_resource_name(:index, model: MyApp.Post, as: :my_posts)
+      :my_posts
+  """
+  def get_resource_name(action, opts) do
+    case opts[:as] do
+      nil ->
+        opts[:model]
+        |> Module.split()
+        |> List.last()
+        |> Macro.underscore()
+        |> pluralize_if_needed(action, opts)
+        |> String.to_atom()
+
+      as ->
+        as
+    end
+  end
+
+  @doc false
+  def persisted?(opts) do
+    !!Keyword.get(opts, :persisted, false) || !!Keyword.get(opts, :required, false)
+  end
+
+  defp pluralize_if_needed(name, action, opts) do
+    if action in [:index] and not persisted?(opts) do
+      name <> "s"
+    else
+      name
+    end
+  end
+
+  @doc """
+  Returns the non-id actions, including any extra actions from `:non_id_actions`.
+  """
+  def non_id_actions(opts) do
+    if opts[:non_id_actions] do
+      Enum.concat([:index, :new, :create], opts[:non_id_actions])
+    else
+      [:index, :new, :create]
+    end
+  end
+
+  @doc """
+  Check if the not_found handler should be applied for the given action, assigns, and options.
+  """
+  @spec apply_handle_not_found?(action :: atom(), assigns :: map(), opts :: Keyword.t()) ::
+          boolean()
+  def apply_handle_not_found?(action, assigns, opts) do
+    non_id_actions = non_id_actions(opts)
+    is_required = required?(opts)
+
+    resource_name = Map.get(assigns, get_resource_name(action, opts))
+
+    if is_nil(resource_name) and (is_required or action not in non_id_actions) do
+      true
+    else
+      false
+    end
+  end
+
+  @doc false
+  def validate_opts(opts) do
+    opts
+    |> warn_deprecated_opts()
+  end
+
+  defp warn_deprecated_opts(opts) do
+    if Keyword.has_key?(opts, :persisted) do
+      IO.warn(
+        "The `:persisted` option is deprecated and will be removed in Canary 2.1.0. Use `:required` instead. Check the documentation for more information."
+      )
+    end
+
+    if Keyword.has_key?(opts, :non_id_actions) do
+      IO.warn(
+        "The `:non_id_actions` option is deprecated and will be removed in Canary 2.1.0. Use a separate :authorize_resource plug for non-id actions and `:except` to skip them on load. Check the documentation for more information."
+      )
+    end
+
+    opts
+  end
 end